+++ system-oceniania/src/main/java/pl/studenci/systemoceniania/repository/UzytkownikRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import pl.studenci.systemoceniania.entity.Uzytkownik;
 import java.util.Optional;
 
-public interface UzytkownikRepository extends JpaRepository<Uzytkownik, Long> {
+public interface UzytkownikRepository extends JpaRepository<Uzytkownik, Integer> {
     Optional<Uzytkownik> findByUsername(String username);
     Optional<Uzytkownik> findByEmail(String email);
 }
